Move playlists components to its own files
- improve readability

[#187173463]
