+++ src/hooks/api/playlistsQuery.test.ts
@@ -16,7 +16,7 @@ import {
 } from 'boclips-api-client/dist/test-support';
 import { VideoFactory } from 'boclips-api-client/dist/test-support/VideosFactory';
 import { CollectionPermission } from 'boclips-api-client/dist/sub-clients/collections/model/CollectionPermissions';
-import { PLAYLISTS_PAGE_SIZE } from 'src/components/playlists/Playlists';
+import { PLAYLISTS_PAGE_SIZE } from 'src/components/playlists/PlaylistList';
 
 describe('playlistsQuery', () => {
   it('will use list projection and convert page size when loading users playlists', async () => {
