+++ src/views/playlist/PlaylistView.tsx
@@ -5,7 +5,7 @@ import { usePlaylistQuery } from 'src/hooks/api/playlistsQuery';
 import { useLocation, useParams } from 'react-router-dom';
 import Footer from 'src/components/layout/Footer';
 import PlaylistHeader from 'src/components/playlists/playlistHeader/PlaylistHeader';
-import PlaylistBody from 'src/components/playlists/PlaylistBody';
+import PlaylistBody from 'src/components/playlists/playlistBody/PlaylistBody';
 import SkeletonPage from 'src/components/skeleton/SkeletonPage';
 import { FollowPlaylist } from 'src/services/followPlaylist';
 import { displayNotification } from 'src/components/common/notification/displayNotification';
